refactor: Move ProviderAPI to `@/lib/providers`
